+++ ush/radmon_verf_bcor.sh
@@ -146,7 +146,7 @@ else
             input_file=${type}
          fi
 
-         rm input
+         if [[ -f input ]]; then rm input; fi
 
       # Check for 0 length input file here and avoid running 
       # the executable if $input_file doesn't exist or is 0 bytes
@@ -200,19 +200,21 @@ EOF
    ${USHradmon}/rstprod.sh
    tar_file=radmon_bcor.tar
 
-   tar -cf $tar_file bcor*.ieee_d* bcor*.ctl*
-   ${COMPRESS} ${tar_file}
-   mv $tar_file.${Z} ${TANKverf_rad}/.
+   if compgen -G "bcor*.ieee_d* bcor*.ctl*" > /dev/null; then
+     tar -cf $tar_file bcor*.ieee_d* bcor*.ctl*
+     ${COMPRESS} ${tar_file}
+     mv $tar_file.${Z} ${TANKverf_rad}/.
 
-   if [[ $RAD_AREA = "rgn" ]]; then
-      cwd=$(pwd)
-      cd ${TANKverf_rad}
-      tar -xf ${tar_file}.${Z}
-      rm ${tar_file}.${Z}
-      cd ${cwd}
+     if [[ $RAD_AREA = "rgn" ]]; then
+        cwd=$(pwd)
+        cd ${TANKverf_rad}
+        tar -xf ${tar_file}.${Z}
+        rm ${tar_file}.${Z}
+        cd ${cwd}
+     fi
    fi
 
-   if [[ $fail -eq $ctr || $fail -gt $ctr ]]; then
+   if [[ $ctr -gt 0 && $fail -eq $ctr || $fail -gt $ctr ]]; then
       err=7
    fi
 fi
